% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{fitdistr}
\alias{fitdistr}
\title{Given a vector of counts, fit a distribution}
\usage{
fitdistr(
  x,
  dist = c("pois", "gamma", "nb", "lnorm", "zipois", "zigamma", "zinb", "zilnorm")
)
}
\arguments{
\item{x}{vector of observed values}

\item{dist}{distribution. Can be chosen from pois, gamma, nb, lnorm, zipois, zigamma, zinb and zilnorm}
}
\description{
Given a vector of counts, fit a distribution
}
