% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{fitdistr_mat}
\alias{fitdistr_mat}
\title{fit distributions to multiple features in a count matrix at the same time}
\usage{
fitdistr_mat(
  count_mat,
  dist = c("pois", "gamma", "nb", "lnorm", "zipois", "zigamma", "zinb", "zilnorm"),
  cutoff = 0.97
)
}
\arguments{
\item{count_mat}{nfeatures*nsamples}

\item{dist}{distribution name among "pois", "gamma", "nb", "lnorm", "zipois", "zigamma", "zinb", "zilnorm"}

\item{cutoff}{values larger than certain quantile are excluded from fitting the model, default 97\%}
}
\description{
fit distributions to multiple features in a count matrix at the same time
}
