% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{gausscopula}
\alias{gausscopula}
\title{simulate from multivariate normal distributions}
\usage{
gausscopula(n = 10, Sigma = diag(nrow = 20))
}
\arguments{
\item{n}{sample size}

\item{Sigma}{covariance matrix whose diagonals are ones. Be default identity matrix}
}
\description{
simulate from multivariate normal distributions
}
