% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_setup}
\alias{graph_setup}
\title{Generate a binary adjacency matrix representing a graph}
\usage{
graph_setup(dimension, sparseness = 0.1)
}
\arguments{
\item{dimension}{an integer representing the number of nodes}

\item{sparseness}{proportion of connected edges among all the pair of nodes}
}
\value{
a binary matrix
}
\description{
Generate a binary adjacency matrix representing a graph
}
