% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gut_experiment_subset}
\alias{gut_experiment_subset}
\title{Stool Metagenomics Data}
\format{
An object of class \code{SummarizedExperiment} with 13794 rows and 1098 columns.
}
\source{
\url{https://doi.org/10.1038/s41591-020-01183-8}
}
\usage{
gut_experiment_subset
}
\description{
A subset of gut microbiome data from the AsnicarF_2021 study retrieved from curatedMetagenomicData
package. The data contains 1098 healthy adult stool samples and 13794 taxa whose prevalence are at least 20\%.
}
\keyword{datasets}
