% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{loglnorm}
\alias{loglnorm}
\title{sum of log likelihoods for log normal distribution}
\usage{
loglnorm(param, x)
}
\arguments{
\item{param}{a named vector with three elements. The first one is the meanlog (logmu) and the other is the sdlog (logsigma)}

\item{x}{vector of observed counts}
}
\description{
sum of log likelihoods for log normal distribution
}
