% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{lognb}
\alias{lognb}
\title{sum of log likelihoods for negative binomial distribution}
\usage{
lognb(param, x)
}
\arguments{
\item{param}{a named vector with two elements. One is the mean (logmu) and the other is the size parameter (logsize)}

\item{x}{vector of observed counts}
}
\description{
sum of log likelihoods for negative binomial distribution
}
