% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{logpois}
\alias{logpois}
\title{sum of log likelihoods for a poisson distribution}
\usage{
logpois(param, x)
}
\arguments{
\item{param}{a named vector with one element "loglambda"}

\item{x}{vector of observed counts}
}
\value{
sum of log likelihoods
}
\description{
sum of log likelihoods for a poisson distribution
}
