% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{logzilnorm}
\alias{logzilnorm}
\title{sum of log likelihoods for zero inflated log normal distribution}
\usage{
logzilnorm(param, x)
}
\arguments{
\item{param}{a named vector with three elements. The first one is the meanlog (mu) and the other is the sdlog (sigma)}

\item{x}{vector of observed counts}
}
\description{
sum of log likelihoods for zero inflated log normal distribution
}
