% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{logzinb}
\alias{logzinb}
\title{sum of log likelihoods for zero inflated negative binomial distribution}
\usage{
logzinb(param, x)
}
\arguments{
\item{param}{a named vector with three elements. The first one is the zero probability (pi0).
The other two are  the mean (mu) and the size parameter (size) of negative binomial distribution.}

\item{x}{vector of observed counts}
}
\description{
sum of log likelihoods for zero inflated negative binomial distribution
}
