% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{logzipois}
\alias{logzipois}
\title{sum of log likelihoods for a zero inflated poisson distribution}
\usage{
logzipois(param, x)
}
\arguments{
\item{param}{a named vector two elements. One is the zero probability (logitpi0) and the other is the mean of nonzero part (loglambda)}

\item{x}{vector of observed counts}
}
\description{
sum of log likelihoods for a zero inflated poisson distribution
}
