% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{mom2param}
\alias{mom2param}
\title{Convert mean and variances to parameters of a distribution}
\usage{
mom2param(
  moments,
  dist = c("pois", "gamma", "nb", "lnorm", "zipois", "zigamma", "zinb", "zilnorm")
)
}
\arguments{
\item{moments}{a named vector with mean, variance and zero inflation probability}

\item{dist}{distribution name, can be one of "pois", "gamma", "nb", "lnorm", "zipois", "zigamma", "zinb", "zilnorm"}
}
\value{
a named vector with the distribution parameters
}
\description{
Convert mean and variances to parameters of a distribution
}
\details{
If the distribution does not have zero inflation components, the names in the vector are "m" and "v".
Otherwise the names need to have "pi0", "mnz" and "vnz". Poisson distribution has mean parameter "lambda".
Gamma distribution has two parameters "shape" and "scale". Negative binomial distribution has two parameters "mu" and "size".
lognormal distribution has two parameters "mu" and "sigma"
}
