% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{param2mom}
\alias{param2mom}
\title{convert parameters to mean and variances for a given distribution}
\usage{
param2mom(
  param,
  dist = c("pois", "gamma", "nb", "lnorm", "zipois", "zigamma", "zinb", "zilnorm")
)
}
\arguments{
\item{param}{a named vector of distribution parameters}

\item{dist}{the name of distribution, can be poisson(pois), gamma(gamma), negative binomial(nb), log normal (lnorm)
and their zero inflated versions(zipois, zigamma, zinb, zilnorm)}
}
\value{
a named vector with mean and variances
}
\description{
convert parameters to mean and variances for a given distribution
}
\details{
For poisson distribution, the name need to be "lambda". For negative binomial distribution,
the name need to be "mu" and "size". For gamma distribution,  the name needs to be "shape" and "scale".
For log normal distribution, the name needs to be "mu" and "sigma". Zero inflated distributions need to have "pi0".
}
