% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{prec2cov}
\alias{prec2cov}
\title{Convert a precision matrix to a covariance/correlation matrix}
\usage{
prec2cov(invsigma, corr = FALSE)
}
\arguments{
\item{invsigma}{a precision matrix that is symmetric and positive definite}

\item{corr}{logical, whether to return a correlation matrix}
}
\description{
Convert a precision matrix to a covariance/correlation matrix
}
