% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{prec_setup}
\alias{prec_setup}
\title{Generate a precision matrix}
\usage{
prec_setup(dimension, sparseness = 0.1, cond = 50, lbound = 1, ubound = 2)
}
\arguments{
\item{dimension}{an integer representing the number of nodes}

\item{sparseness}{proportion of connected edges among all the pair of nodes}

\item{cond}{target condition number, must be larger than one}

\item{lbound}{lower bound of absolute nonzero value for off diagonal terms}

\item{ubound}{upper bound of absolute nonzero value for off diagonal terms}
}
\description{
Generate a positive definite matrix with given sparseness of dependencies and condition number
}
