% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simcount}
\alias{simcount}
\title{generate counts based on the given distribution and the quantiles}
\usage{
simcount(
  q,
  param,
  dist = c("pois", "gamma", "nb", "lnorm", "zipois", "zigamma", "zinb", "zilnorm")
)
}
\arguments{
\item{q}{vector of quantiles}

\item{param}{distribution parameters. A named vector}

\item{dist}{distribution name. Choose between "pois", "gamma", "nb", "lnorm", "zipois", "zigamma", "zinb", "zilnorm"}
}
\description{
generate counts based on the given distribution and the quantiles
}
