% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simcountmat}
\alias{simcountmat}
\title{simulate count matrix given the quantiles and the distribution type and parameters.}
\usage{
simcountmat(
  copula,
  params,
  dist = c("pois", "gamma", "nb", "lnorm", "zipois", "zigamma", "zinb", "zilnorm")
)
}
\arguments{
\item{copula}{a matrix of quantile values, nsample * nfeature}

\item{params}{distribution parameters for each feature}

\item{dist}{distributions}
}
\description{
simulate count matrix given the quantiles and the distribution type and parameters.
}
