% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{loggamma}
\alias{loggamma}
\title{sum of log likelihoods for a gamma distribution}
\usage{
loggamma(param, x)
}
\arguments{
\item{param}{a named vector two elements. One is the shape parameter (logalpha) and the other is the scale (logbeta)}

\item{x}{vector of observed counts}
}
\description{
sum of log likelihoods for a gamma distribution
}
