% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{logzigamma}
\alias{logzigamma}
\title{sum of log likelihoods for a zero inflated gamma distribution}
\usage{
logzigamma(param, x)
}
\arguments{
\item{param}{a named vector with three elements. One is the zero probability pi0.
The other two are shape parameter (alpha) and the other is the scale (beta)}

\item{x}{vector of observed counts}
}
\description{
sum of log likelihoods for a zero inflated gamma distribution
}
